Major changes to remove Jackson dependencies etc.
