+++ criteria/sql/src/org/immutables/criteria/sql/SqlException.java
