+++ criteria/sql/src/org/immutables/criteria/sql/commands/SqlDeleteCommand.java
