+++ criteria/sql/src/org/immutables/criteria/sql/compiler/SqlConstantExpression.java
