+++ criteria/sql/src/org/immutables/criteria/sql/compiler/SqlPathNaming.java
