+++ criteria/sql/src/org/immutables/criteria/sql/conversion/TypeConverter.java
@@ -17,5 +17,5 @@ package org.immutables.criteria.sql.conversion;
 
 @FunctionalInterface
 public interface TypeConverter<F, T> {
-    T apply(final F value);
+  T apply(final F value);
 }
