+++ criteria/sql/src/org/immutables/criteria/sql/reflection/PropertyExtractor.java
@@ -17,5 +17,5 @@ package org.immutables.criteria.sql.reflection;
 
 @FunctionalInterface
 public interface PropertyExtractor {
-    Object extract(Object entity);
+  Object extract(Object entity);
 }
