+++ criteria/sql/test/org/immutables/criteria/sql/compiler/Dummy.java
@@ -21,8 +21,8 @@ import org.immutables.value.Value;
 @Criteria
 @Value.Immutable
 public interface Dummy {
-    @Criteria.Id
-    int id();
+  @Criteria.Id
+  int id();
 
-    String name();
+  String name();
 }
